pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.2.0",
    "ffmpeg-python>=0.2.0",
    "gtts>=2.5.4",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "moviepy==1.0.3",
    "pydub>=0.25.1",
    "pysrt>=1.1.2",
    "streamlit>=1.50.0",
    "translate>=3.6.1",
]
